Initial C-style interface for whisper.cpp
